+++ src/main/java/appeng/core/Registration.java
@@ -64,6 +64,7 @@ import appeng.core.features.registries.cell.BasicCellHandler;
 import appeng.core.features.registries.cell.CreativeCellHandler;
 import appeng.parts.automation.InitStackWorldBehaviors;
 import appeng.parts.misc.InitExternalStorageStrategies;
+import appeng.util.inv.InitGenericInventoryAdapters;
 import appeng.core.localization.GuiText;
 import appeng.core.sync.ChannelTierSync;
 import appeng.core.sync.MultiblockLimitSync;
@@ -407,6 +408,7 @@ final class Registration {
         // which happens in newRegistry() (RegistryEvent.NewRegistry) - long before this FMLInitializationEvent.
         InitStackWorldBehaviors.register();
         InitExternalStorageStrategies.register();
+        InitGenericInventoryAdapters.register();
 
         api.definitions().materials().matterBall().maybeStack(1).ifPresent(ammoStack ->
         {